refactor(cli): extract profile + backend-native groups, diag cluster (3a slice 3)
Completes the 3a CLI group extractions. Pulls the last two command groups
(profile, backend-native) into cli_profile.py / cli_backend_native.py, and
lifts the shared diagnostic helper cluster (DB-health probes, profile
reporters, doctor report builder, recall-daemon health) into a new
cli_diag.py.

- cli_diag.py: _managed_sqlite_dbs, _sqlite_* probes, _db_health_report,
  _profile_active/expected_config, _profile_overrides, _model_cache_report,
  _typed_embedder_profile, _profile_status_report, _profile_repair_plan,
  _json_import_check, _doctor_report, _recall_daemon_health. Pure cfg->dict
  reporters shared by doctor/stats (still in cli.py) and the profile +
  backend-native groups. _doctor_report defers its import of
  _runtime_install_report (stays in cli.py) to avoid a cli<->cli_diag cycle.
- _backend_native_trace_id + _memo_backend_version -> cli_common (needed by
  the extracted backend-native group; no longer referenced in cli.py).
- Dropped the now-orphan _now_iso from cli.py.
- Updated tests/test_profile.py + tests/test_embedder_profile_authority.py
  to import the relocated helpers from memo.cli_diag.

cli.py is now ~5180 LOC (from 9138 at the start of 3a). Verification: ruff
clean, mypy clean (86 files), pytest 736 passed, --help smoke on profile /
backend-native / doctor.
